data_pipeline/requirements-build.txt
<<<<<<< HEAD
--extra-index-url https://download.pytorch.org/whl/cu118

datetime
k-means-constrained
numpy
python-dotenv
psycopg2-binary
torch
torchvision
torchaudio
transformers
tqdm
=======
langchain_core==0.3.61
langchain_openai==0.3.18
langgraph==0.4.7
psycopg2_binary==2.9.10
pydantic==2.11.5
python-dotenv==1.1.0
Requests==2.32.3
tavily-python==0.7.3
tqdm==4.66.4
typing_extensions==4.13.2
urllib3==2.4.0
>>>>>>> agentic_classification
